move public types to interfaces
