+++ src/client/grants.rs
@@ -4,12 +4,12 @@ use anyhow::{Result, anyhow};
 use base64ct::{Base64UrlUnpadded, Encoding};
 use chrono::{DateTime, Duration, Utc};
 
+use crate::client::records::{Data, ProtocolBuilder, WriteBuilder};
 pub use crate::grants::{Conditions, Publication, RecordsScope, Scope};
 use crate::grants::{GrantData, RequestData, RevocationData};
-use crate::interfaces::records::{Data, ProtocolBuilder, WriteBuilder};
+use crate::interfaces::records::{self, Tag, Write};
 use crate::protocols::{self};
 use crate::provider::Signer;
-use crate::records::{self, Tag, Write};
 use crate::{Interface, utils};
 
 /// Options to use when creating a permission grant.
