+++ src/interfaces/records.rs
@@ -1,1179 +1,1032 @@
-//! # Records Interface
+//! # Records
 //!
-//! The Records interface provides a mechanism to store data using shared
-//! schemas.
+//! The records delete endpoint handles `RecordsDelete` messages — requests
+//! to delete a [`Write`] record.
 //!
-//! Shared schemas, some of which may be well-known for a given domain, allow
-//! DWN-baswed apps and services to share datasets with one another. This leads
-//! to improved cross-app experiences for users.
+//! Technically, the [`Write`] record is not deleted, but rather a new
+//! [`Delete`] record is created to mark the record as deleted. The [`Delete`]
+//! record is used to prune the record and its descendants from the system,
+//! leaving only the [`Delete`] and initial [`Write`] records.
 
-use std::collections::HashMap;
-use std::io::Cursor;
+use std::collections::{BTreeMap, HashMap};
+use std::fmt::Display;
+use std::io;
 
-use anyhow::{Result, anyhow};
+use base64ct::{Base64UrlUnpadded, Encoding};
 use chrono::{DateTime, Utc};
-use credibil_infosec::jose::{Jws, JwsBuilder};
+use credibil_infosec::jose::jwe::{self, ContentAlgorithm, KeyAlgorithm, Protected};
+use credibil_infosec::jose::{Curve, Jws, PublicKeyJwk};
+use serde::{Deserialize, Serialize};
+use zeroize::{Zeroize, ZeroizeOnDrop};
 
-use crate::authorization::{Authorization, AuthorizationBuilder};
+use super::{Cursor, DateRange, Descriptor, Pagination, Range};
+use crate::authorization::{Authorization, JwsPayload};
+use crate::event::Subscriber;
 use crate::hd_key::DerivationScheme;
-use crate::provider::Signer;
-pub use crate::records::{
-    Attestation, DelegatedGrant, DeleteDescriptor, EncryptOptions, Recipient, RecordsFilter,
-    SignaturePayload, Sort, decrypt,
-};
-use crate::records::{
-    Delete, EncryptionProperty, Query, QueryDescriptor, Read, ReadDescriptor, Subscribe,
-    SubscribeDescriptor, Tag, Write, WriteDescriptor,
-};
-use crate::store::Pagination;
-use crate::utils::cid;
-use crate::{Descriptor, Interface, Method, utils};
+use crate::serde::{rfc3339_micros, rfc3339_micros_opt};
+use crate::{OneOrMany, Result, unexpected, utils};
 
-/// Options to use when creating a permission grant.
-pub struct DeleteBuilder<R, S> {
-    message_timestamp: DateTime<Utc>,
-    record_id: R,
-    prune: Option<bool>,
-    permission_grant_id: Option<String>,
-    protocol_role: Option<String>,
-    signer: S,
+/// The [`Delete`] message expected by the handler.
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct Delete {
+    /// Delete descriptor.
+    pub descriptor: DeleteDescriptor,
+
+    /// Message authorization.
+    pub authorization: Authorization,
 }
 
-/// Builder state is unfiltered.
-#[doc(hidden)]
-pub struct NoRecordId;
-/// Builder state is filtered.
-#[doc(hidden)]
-pub struct RecordId(String);
+/// The [`Delete`] message descriptor.
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct DeleteDescriptor {
+    /// The base descriptor
+    #[serde(flatten)]
+    pub base: Descriptor,
 
-impl Default for DeleteBuilder<NoRecordId, Unsigned> {
-    fn default() -> Self {
-        Self::new()
-    }
+    /// The ID of the record to delete.
+    pub record_id: String,
+
+    /// Specifies whether descendent records should be pruned or not.
+    pub prune: bool,
 }
 
-impl DeleteBuilder<NoRecordId, Unsigned> {
-    /// Returns a new [`DeleteBuilder`]
+/// [`DeleteReply`] is returned by the handler in the [`Reply`] `body` field.
+#[derive(Debug)]
+pub struct DeleteReply;
+
+/// The [`Query`] message expected by the handler.
+#[derive(Clone, Debug, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct Query {
+    /// The Query descriptor.
+    pub descriptor: QueryDescriptor,
+
+    /// The message authorization.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub authorization: Option<Authorization>,
+}
+
+/// The [`Query`] message descriptor.
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct QueryDescriptor {
+    /// The base descriptor
+    #[serde(flatten)]
+    pub base: Descriptor,
+
+    /// Filter Records for query.
+    pub filter: RecordsFilter,
+
+    /// Specifies how dates should be sorted.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub date_sort: Option<Sort>,
+
+    /// The pagination cursor.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub pagination: Option<Pagination>,
+}
+
+/// [`QueryReply`] is returned by the handler in the [`Reply`] `body` field.
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct QueryReply {
+    /// Query reply entries.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub entries: Option<Vec<QueryReplyEntry>>,
+
+    /// Pagination cursor.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub cursor: Option<Cursor>,
+}
+
+/// [`QueryReplyEntry`] represents a [`Write`] entry returned by the query.
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct QueryReplyEntry {
+    /// The `RecordsWrite` message of the record if record exists.
+    #[serde(flatten)]
+    pub write: Write,
+
+    /// The initial write of the record if the returned `RecordsWrite` message
+    /// itself is not the initial write.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub initial_write: Option<Write>,
+}
+
+/// The Records filter is used when querying for records.
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct RecordsFilter {
+    /// Find a single record by its ID. May return two results — an initial
+    /// write and the latest update or Delete.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub record_id: Option<String>,
+
+    /// Records matching the specified author.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub author: Option<OneOrMany<String>>,
+
+    /// Records matching the specified creator.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub attester: Option<String>,
+
+    /// Records matching the specified recipient(s).
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub recipient: Option<OneOrMany<String>>,
+
+    /// Records with the specified context.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub context_id: Option<String>,
+
+    /// The CID of the parent object .
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub parent_id: Option<String>,
+
+    /// Entry matching the specified protocol.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub protocol: Option<String>,
+
+    /// Entry protocol path.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub protocol_path: Option<String>,
+
+    /// Records with the specified schema.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub schema: Option<String>,
+
+    /// The MIME type of the requested data. For example, `application/json`.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub data_format: Option<String>,
+
+    /// Match records with the specified tags.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub tags: Option<BTreeMap<String, TagFilter>>,
+
+    /// CID of the data.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub data_cid: Option<String>,
+
+    /// Records with a size within the range.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub data_size: Option<Range<usize>>,
+
+    /// Whether the record is published.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub published: Option<bool>,
+
+    /// Filter messages published within the specified range.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub date_published: Option<DateRange>,
+
+    /// Filter messages created within the specified range.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub date_created: Option<DateRange>,
+
+    /// Match messages updated within the specified range.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub date_updated: Option<DateRange>,
+}
+
+impl RecordsFilter {
+    /// Create a new [`RecordsFilter`]
     #[must_use]
     pub fn new() -> Self {
-        Self {
-            message_timestamp: Utc::now(),
-            record_id: NoRecordId,
-            prune: None,
-            permission_grant_id: None,
-            protocol_role: None,
-            signer: Unsigned,
-        }
+        Self::default()
     }
 
-    /// Specifies the permission grant ID.
-    #[must_use]
-    pub fn record_id(self, record_id: impl Into<String>) -> DeleteBuilder<RecordId, Unsigned> {
-        DeleteBuilder {
-            record_id: RecordId(record_id.into()),
-
-            message_timestamp: self.message_timestamp,
-            prune: self.prune,
-            permission_grant_id: self.permission_grant_id,
-            protocol_role: self.protocol_role,
-            signer: self.signer,
-        }
-    }
-}
-
-impl<R> DeleteBuilder<R, Unsigned> {
-    /// Specifies the permission grant ID.
-    #[must_use]
-    pub const fn prune(mut self, prune: bool) -> Self {
-        self.prune = Some(prune);
-        self
-    }
-
-    /// Specifies the permission grant ID.
-    #[must_use]
-    pub fn permission_grant_id(mut self, permission_grant_id: impl Into<String>) -> Self {
-        self.permission_grant_id = Some(permission_grant_id.into());
-        self
-    }
-
-    /// Specifies the permission grant ID.
-    #[must_use]
-    pub fn protocol_role(mut self, protocol_role: impl Into<String>) -> Self {
-        self.protocol_role = Some(protocol_role.into());
-        self
-    }
-}
-
-impl DeleteBuilder<RecordId, Unsigned> {
-    /// Logically (from user POV), sign the record.
-    ///
-    /// At this point, the builder simply captures the signer for use in the
-    /// final build step.
-    #[must_use]
-    pub fn sign<S: Signer>(self, signer: &S) -> DeleteBuilder<RecordId, Signed<'_, S>> {
-        DeleteBuilder {
-            signer: Signed(signer),
-
-            message_timestamp: self.message_timestamp,
-            record_id: self.record_id,
-            prune: self.prune,
-            permission_grant_id: self.permission_grant_id,
-            protocol_role: self.protocol_role,
-        }
-    }
-}
-
-impl<S: Signer> DeleteBuilder<RecordId, Signed<'_, S>> {
-    /// Build the write message.
-    ///
-    /// # Errors
-    ///
-    /// This method will fail when there is an issue authorizing the message.
-    pub async fn build(self) -> Result<Delete> {
-        let descriptor = DeleteDescriptor {
-            base: Descriptor {
-                interface: Interface::Records,
-                method: Method::Delete,
-                message_timestamp: self.message_timestamp,
-            },
-            record_id: self.record_id.0,
-            prune: self.prune.unwrap_or(false),
-        };
-
-        let mut auth_builder =
-            AuthorizationBuilder::new().descriptor_cid(cid::from_value(&descriptor)?);
-        if let Some(id) = self.permission_grant_id {
-            auth_builder = auth_builder.permission_grant_id(id);
-        }
-        if let Some(role) = self.protocol_role {
-            auth_builder = auth_builder.protocol_role(role);
-        }
-        let authorization = auth_builder.build(self.signer.0).await?;
-
-        Ok(Delete {
-            descriptor,
-            authorization,
-        })
-    }
-}
-
-/// Options to use when creating a permission grant.
-pub struct QueryBuilder<F, S> {
-    message_timestamp: DateTime<Utc>,
-    filter: F,
-    date_sort: Option<Sort>,
-    pagination: Option<Pagination>,
-    protocol_role: Option<String>,
-    permission_grant_id: Option<String>,
-    delegated_grant: Option<DelegatedGrant>,
-    signer: S,
-}
-
-/// Builder state is unsigned.
-#[doc(hidden)]
-pub struct Unsigned;
-/// Builder state is signed.
-#[doc(hidden)]
-pub struct Signed<'a, S: Signer>(pub &'a S);
-
-/// Builder state is unfiltered.
-#[doc(hidden)]
-pub struct Unfiltered;
-/// Builder state is filtered.
-#[doc(hidden)]
-pub struct Filtered(RecordsFilter);
-
-impl Default for QueryBuilder<Unfiltered, Unsigned> {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl QueryBuilder<Unfiltered, Unsigned> {
-    /// Returns a new [`QueryBuilder`]
-    #[must_use]
-    pub fn new() -> Self {
-        Self {
-            message_timestamp: Utc::now(),
-            filter: Unfiltered,
-            signer: Unsigned,
-            date_sort: None,
-            pagination: None,
-            protocol_role: None,
-            permission_grant_id: None,
-            delegated_grant: None,
-        }
-    }
-
-    /// Set the filter to use when querying.
-    #[must_use]
-    pub fn filter(self, filter: RecordsFilter) -> QueryBuilder<Filtered, Unsigned> {
-        QueryBuilder {
-            filter: Filtered(filter),
-            message_timestamp: self.message_timestamp,
-            date_sort: self.date_sort,
-            pagination: self.pagination,
-            signer: self.signer,
-            protocol_role: self.protocol_role,
-            permission_grant_id: self.permission_grant_id,
-            delegated_grant: self.delegated_grant,
-        }
-    }
-}
-
-/// State: Unsigned
-impl<'a, F> QueryBuilder<F, Unsigned> {
-    /// Specifies the permission grant ID.
-    #[must_use]
-    pub fn permission_grant_id(mut self, permission_grant_id: impl Into<String>) -> Self {
-        self.permission_grant_id = Some(permission_grant_id.into());
-        self
-    }
-
-    /// Specify a protocol role for the record.
-    #[must_use]
-    pub fn protocol_role(mut self, protocol_role: impl Into<String>) -> Self {
-        self.protocol_role = Some(protocol_role.into());
-        self
-    }
-
-    /// The delegated grant used with this record.
-    #[must_use]
-    pub fn delegated_grant(mut self, delegated_grant: DelegatedGrant) -> Self {
-        self.delegated_grant = Some(delegated_grant);
-        self
-    }
-
-    /// Determines which date to use when sorting query results.
-    #[must_use]
-    pub const fn date_sort(mut self, date_sort: Sort) -> Self {
-        self.date_sort = Some(date_sort);
-        self
-    }
-
-    /// Sets the limit (size) and offset of the resultset pagination cursor.
-    #[must_use]
-    pub fn pagination(mut self, pagination: Pagination) -> Self {
-        self.pagination = Some(pagination);
-        self
-    }
-
-    /// Logically (from user POV), sign the record.
-    ///
-    /// At this point, the builder simply captures the signer for use in the
-    /// final build step.
-    #[must_use]
-    pub fn sign<S: Signer>(self, signer: &'a S) -> QueryBuilder<F, Signed<'a, S>> {
-        QueryBuilder {
-            signer: Signed(signer),
-
-            message_timestamp: self.message_timestamp,
-            filter: self.filter,
-            date_sort: self.date_sort,
-            pagination: self.pagination,
-            protocol_role: self.protocol_role,
-            permission_grant_id: self.permission_grant_id,
-            delegated_grant: self.delegated_grant,
-        }
-    }
-}
-
-// Build without signing
-impl QueryBuilder<Filtered, Unsigned> {
-    /// Build the write message.
-    ///
-    /// # Errors
-    ///
-    /// This method will fail when there is an issue normalizing filter URLs.
-    pub fn build(self) -> Result<Query> {
-        Ok(Query {
-            descriptor: QueryDescriptor {
-                base: Descriptor {
-                    interface: Interface::Records,
-                    method: Method::Query,
-                    message_timestamp: self.message_timestamp,
-                },
-                filter: self.filter.0.normalize()?,
-                date_sort: self.date_sort,
-                pagination: self.pagination,
-            },
-            authorization: None,
-        })
-    }
-}
-
-// Build includes signing
-impl<S: Signer> QueryBuilder<Filtered, Signed<'_, S>> {
-    /// Build the write message.
-    ///
-    /// # Errors
-    ///
-    /// This method will fail when there is an issue nomalizng filter URLs or
-    /// authorizing the message.
-    pub async fn build(self) -> Result<Query> {
-        let descriptor = QueryDescriptor {
-            base: Descriptor {
-                interface: Interface::Records,
-                method: Method::Query,
-                message_timestamp: self.message_timestamp,
-            },
-            filter: self.filter.0.normalize()?,
-            date_sort: self.date_sort,
-            pagination: self.pagination,
-        };
-
-        let mut auth_builder =
-            AuthorizationBuilder::new().descriptor_cid(cid::from_value(&descriptor)?);
-        if let Some(id) = self.permission_grant_id {
-            auth_builder = auth_builder.permission_grant_id(id);
-        }
-        if let Some(role) = self.protocol_role {
-            auth_builder = auth_builder.protocol_role(role);
-        }
-        if let Some(delegated_grant) = self.delegated_grant {
-            auth_builder = auth_builder.delegated_grant(delegated_grant);
-        }
-        let authorization = Some(auth_builder.build(self.signer.0).await?);
-
-        Ok(Query {
-            descriptor,
-            authorization,
-        })
-    }
-}
-
-/// Options to use when creating a permission grant.
-#[derive(Clone, Debug, Default)]
-pub struct ReadBuilder<F, S> {
-    message_timestamp: DateTime<Utc>,
-    filter: F,
-    permission_grant_id: Option<String>,
-    protocol_role: Option<String>,
-    delegated_grant: Option<DelegatedGrant>,
-    signer: S,
-}
-
-impl Default for ReadBuilder<Unfiltered, Unsigned> {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl ReadBuilder<Unfiltered, Unsigned> {
-    /// Returns a new [`ReadBuilder`]
-    #[must_use]
-    pub fn new() -> Self {
-        Self {
-            message_timestamp: Utc::now(),
-            filter: Unfiltered,
-            permission_grant_id: None,
-            protocol_role: None,
-            delegated_grant: None,
-            signer: Unsigned,
-        }
-    }
-
-    /// Specifies the permission grant ID.
-    #[must_use]
-    pub fn filter(self, filter: RecordsFilter) -> ReadBuilder<Filtered, Unsigned> {
-        ReadBuilder {
-            message_timestamp: self.message_timestamp,
-            filter: Filtered(filter),
-            permission_grant_id: self.permission_grant_id,
-            protocol_role: self.protocol_role,
-            delegated_grant: self.delegated_grant,
-            signer: Unsigned,
-        }
-    }
-}
-
-impl<'a, F> ReadBuilder<F, Unsigned> {
-    /// Specifies the permission grant ID.
-    #[must_use]
-    pub fn permission_grant_id(mut self, permission_grant_id: impl Into<String>) -> Self {
-        self.permission_grant_id = Some(permission_grant_id.into());
-        self
-    }
-
-    // /// Specify a protocol role for the record.
-    // #[must_use]
-    // pub const fn authorize(mut self, authorize: bool) -> Self {
-    //     self.authorize = Some(authorize);
-    //     self
-    // }
-
-    /// Specify a protocol role for the record.
-    #[must_use]
-    pub fn protocol_role(mut self, protocol_role: impl Into<String>) -> Self {
-        self.protocol_role = Some(protocol_role.into());
-        self
-    }
-
-    /// The delegated grant used with this record.
-    #[must_use]
-    pub fn delegated_grant(mut self, delegated_grant: DelegatedGrant) -> Self {
-        self.delegated_grant = Some(delegated_grant);
-        self
-    }
-
-    /// Logically (from user POV), sign the record.
-    ///
-    /// At this point, the builder simply captures the signer for use in the
-    /// final build step.
-    #[must_use]
-    pub fn sign<S: Signer>(self, signer: &'a S) -> ReadBuilder<F, Signed<'a, S>> {
-        ReadBuilder {
-            message_timestamp: self.message_timestamp,
-            filter: self.filter,
-            permission_grant_id: self.permission_grant_id,
-            protocol_role: self.protocol_role,
-            delegated_grant: self.delegated_grant,
-            signer: Signed(signer),
-        }
-    }
-}
-
-impl ReadBuilder<Filtered, Unsigned> {
-    /// Build and return an anonymous (unsigned) Read message.
-    #[must_use]
-    pub fn build(self) -> Read {
-        let descriptor = ReadDescriptor {
-            base: Descriptor {
-                interface: Interface::Records,
-                method: Method::Read,
-                message_timestamp: self.message_timestamp,
-            },
-            filter: self.filter.0,
-        };
-
-        Read {
-            descriptor,
-            authorization: None,
-        }
-    }
-}
-
-impl<S: Signer> ReadBuilder<Filtered, Signed<'_, S>> {
-    /// Build the write message.
-    ///
-    /// # Errors
-    ///
-    /// This method will fail when there is an issue nomalizng filter URLs or
-    /// authorizing the message.
-    pub async fn build(self) -> Result<Read> {
-        let descriptor = ReadDescriptor {
-            base: Descriptor {
-                interface: Interface::Records,
-                method: Method::Read,
-                message_timestamp: self.message_timestamp,
-            },
-            filter: self.filter.0.normalize()?,
-        };
-
-        let mut auth_builder =
-            AuthorizationBuilder::new().descriptor_cid(cid::from_value(&descriptor)?);
-        if let Some(id) = self.permission_grant_id {
-            auth_builder = auth_builder.permission_grant_id(id);
-        }
-        if let Some(role) = self.protocol_role {
-            auth_builder = auth_builder.protocol_role(role);
-        }
-        if let Some(delegated_grant) = self.delegated_grant {
-            auth_builder = auth_builder.delegated_grant(delegated_grant);
-        }
-
-        Ok(Read {
-            descriptor,
-            authorization: Some(auth_builder.build(self.signer.0).await?),
-        })
-    }
-}
-
-/// Options to use when creating a permission grant.
-pub struct SubscribeBuilder<F, S> {
-    message_timestamp: DateTime<Utc>,
-    filter: F,
-    permission_grant_id: Option<String>,
-    protocol_role: Option<String>,
-    delegated_grant: Option<DelegatedGrant>,
-    authorize: Option<bool>,
-    signer: S,
-}
-
-impl Default for SubscribeBuilder<Unfiltered, Unsigned> {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl SubscribeBuilder<Unfiltered, Unsigned> {
-    /// Returns a new [`SubscribeBuilder`]
-    #[must_use]
-    pub fn new() -> Self {
-        Self {
-            message_timestamp: Utc::now(),
-            filter: Unfiltered,
-            permission_grant_id: None,
-            protocol_role: None,
-            delegated_grant: None,
-            authorize: None,
-            signer: Unsigned,
-        }
-    }
-
-    /// Set the filter to use when querying.
-    #[must_use]
-    pub fn filter(self, filter: RecordsFilter) -> SubscribeBuilder<Filtered, Unsigned> {
-        SubscribeBuilder {
-            filter: Filtered(filter),
-            message_timestamp: self.message_timestamp,
-            signer: self.signer,
-            permission_grant_id: self.permission_grant_id,
-            protocol_role: self.protocol_role,
-            delegated_grant: self.delegated_grant,
-            authorize: self.authorize,
-        }
-    }
-}
-
-/// State: Unsigned
-impl<'a, F> SubscribeBuilder<F, Unsigned> {
-    /// Specifies the permission grant ID.
-    #[must_use]
-    pub fn permission_grant_id(mut self, permission_grant_id: impl Into<String>) -> Self {
-        self.permission_grant_id = Some(permission_grant_id.into());
-        self
-    }
-
-    /// Specify a protocol role for the record.
-    #[must_use]
-    pub const fn authorize(mut self, authorize: bool) -> Self {
-        self.authorize = Some(authorize);
-        self
-    }
-
-    /// Specify a protocol role for the record.
-    #[must_use]
-    pub fn protocol_role(mut self, protocol_role: impl Into<String>) -> Self {
-        self.protocol_role = Some(protocol_role.into());
-        self
-    }
-
-    /// The delegated grant used with this record.
-    #[must_use]
-    pub fn delegated_grant(mut self, delegated_grant: DelegatedGrant) -> Self {
-        self.delegated_grant = Some(delegated_grant);
-        self
-    }
-
-    /// Logically (from user POV), sign the record.
-    ///
-    /// At this point, the builder simply captures the signer for use in the
-    /// final build step.
-    #[must_use]
-    pub fn sign<S: Signer>(self, signer: &'a S) -> SubscribeBuilder<F, Signed<'a, S>> {
-        SubscribeBuilder {
-            signer: Signed(signer),
-
-            message_timestamp: self.message_timestamp,
-            filter: self.filter,
-            permission_grant_id: self.permission_grant_id,
-            protocol_role: self.protocol_role,
-            delegated_grant: self.delegated_grant,
-            authorize: self.authorize,
-        }
-    }
-}
-
-// State: Signer set
-impl<S: Signer> SubscribeBuilder<Filtered, Signed<'_, S>> {
-    /// Build the write message.
-    ///
-    /// # Errors
-    ///
-    /// This method will fail when there is an issue nomalizng filter URLs or
-    /// authorizing the message.
-    pub async fn build(self) -> Result<Subscribe> {
-        let descriptor = SubscribeDescriptor {
-            base: Descriptor {
-                interface: Interface::Records,
-                method: Method::Subscribe,
-                message_timestamp: self.message_timestamp,
-            },
-            filter: self.filter.0.normalize()?,
-        };
-
-        let authorization = if self.authorize.unwrap_or(true) {
-            let mut auth_builder =
-                AuthorizationBuilder::new().descriptor_cid(cid::from_value(&descriptor)?);
-            if let Some(id) = self.permission_grant_id {
-                auth_builder = auth_builder.permission_grant_id(id);
-            }
-            if let Some(role) = self.protocol_role {
-                auth_builder = auth_builder.protocol_role(role);
-            }
-            if let Some(delegated_grant) = self.delegated_grant {
-                auth_builder = auth_builder.delegated_grant(delegated_grant);
-            }
-            Some(auth_builder.build(self.signer.0).await?)
+    /// Normalizes protocol and schema URLs within the `RecordsFilter`.
+    pub(crate) fn normalize(&self) -> Result<Self> {
+        let mut filter = self.clone();
+        filter.protocol = if let Some(protocol) = &self.protocol {
+            Some(utils::uri::clean(protocol)?)
         } else {
             None
         };
-
-        Ok(Subscribe {
-            descriptor,
-            authorization,
-        })
+        filter.schema =
+            if let Some(schema) = &self.schema { Some(utils::uri::clean(schema)?) } else { None };
+        Ok(filter)
     }
-}
 
-/// Options for use when creating a new [`Write`] message.
-pub struct WriteBuilder<'a, O, A, S> {
-    message_timestamp: DateTime<Utc>,
-    recipient: Option<String>,
-    protocol: Option<ProtocolBuilder<'a>>,
-    schema: Option<String>,
-    tags: Option<HashMap<String, Tag>>,
-    record_id: Option<String>,
-    data: Option<Data>,
-    data_format: String,
-    date_created: DateTime<Utc>,
-    published: Option<bool>,
-    date_published: Option<DateTime<Utc>>,
-    protocol_role: Option<String>,
-    permission_grant_id: Option<String>,
-    delegated_grant: Option<DelegatedGrant>,
-    existing: Option<Write>,
-    encryption: Option<EncryptionProperty>,
-    origin: O,
-    attesters: A,
-    signer: S,
-}
-
-impl Default for WriteBuilder<'_, New, Unattested, Unsigned> {
-    fn default() -> Self {
-        Self::new()
+    /// Check whether the filter will return a concise set of results.
+    pub(crate) const fn is_concise(&self) -> bool {
+        self.record_id.is_some()
+            || self.protocol_path.is_some()
+            || self.context_id.is_some()
+            || self.parent_id.is_some()
+            || self.schema.is_some()
     }
-}
 
-/// The protocol to use for the Write message.
-#[derive(Clone, Debug, Default)]
-pub struct ProtocolBuilder<'a> {
-    /// Entry protocol.
-    pub protocol: &'a str,
-
-    /// Protocol path.
-    pub protocol_path: &'a str,
-
-    /// Parent context for the protocol.
-    pub parent_context_id: Option<String>,
-}
-
-/// Entry data can be raw bytes or CID.
-pub enum Data {
-    /// Data is a `Cursor`.
-    Stream(Cursor<Vec<u8>>),
-
-    /// Data is use to calculate CID and size of previously stored data — as
-    /// for `Data::Cid`. The data is not added to the Write record's
-    /// `data_stream`.
-    ///
-    /// N.B. This option can only be used when the referenced data has already
-    /// been stored by the web node.
-    Bytes(Vec<u8>),
-
-    /// A CID (and size) referencing `BlockStore` data from a previous update
-    /// to the Write record.
-    ///
-    /// N.B. This option can only be used when the referenced data has already
-    /// been stored by the web node.
-    Cid {
-        /// CID of data already stored by the web node. If not set, the `data`
-        /// parameter must be set.
-        data_cid: String,
-
-        /// Size of the `data` attribute in bytes. Must be set when `data_cid` is set,
-        /// otherwise should be left unset.
-        data_size: usize,
-    },
-}
-
-impl Default for Data {
-    fn default() -> Self {
-        Self::Stream(Cursor::default())
-    }
-}
-
-impl From<Vec<u8>> for Data {
-    fn from(data: Vec<u8>) -> Self {
-        Self::Stream(Cursor::new(data))
-    }
-}
-
-// State 'guards' for the WriteBuilder typestate pattern.
-
-/// The WriteBuilder is in a new state.
-#[doc(hidden)]
-pub struct New;
-/// The WriteBuilder is in an 'existing' state.
-#[doc(hidden)]
-pub struct Existing;
-
-/// The WriteBuilder is in an 'unattested' state.
-#[doc(hidden)]
-pub struct Unattested;
-/// The WriteBuilder is in an 'attested' state.
-#[doc(hidden)]
-pub struct Attested<'a, A: Signer>(pub &'a [&'a A]);
-
-/// Create a `Write` record from scratch.
-impl WriteBuilder<'_, New, Unattested, Unsigned> {
-    /// Returns a new [`WriteBuilder`]
-    #[must_use]
-    pub fn new() -> Self {
-        let now = Utc::now();
-
-        Self {
-            message_timestamp: now,
-            date_created: now,
-            data: None,
-            data_format: "application/json".to_string(),
-            signer: Unsigned,
-            attesters: Unattested,
-            origin: New,
-            recipient: None,
-            protocol: None,
-            schema: None,
-            tags: None,
-            record_id: None,
-            published: None,
-            date_published: None,
-            protocol_role: None,
-            permission_grant_id: None,
-            delegated_grant: None,
-            existing: None,
-            encryption: None,
+    /// Create an optimized filter to use with single-field indexes. This
+    /// method chooses the best filter property, in order of priority, to use
+    /// when querying.
+    pub(crate) fn as_concise(&self) -> Option<(String, String)> {
+        if let Some(record_id) = &self.record_id {
+            return Some(("recordId".to_string(), record_id.clone()));
         }
-    }
-}
-
-/// Create a [`Write`] record from an existing record.
-impl WriteBuilder<'_, Existing, Unattested, Unsigned> {
-    /// Returns a new [`WriteBuilder`] based on an existing `Write` record.
-    #[must_use]
-    pub fn from(existing: Write) -> Self {
-        let mut existing = existing;
-        existing.data_stream = None;
-        existing.encoded_data = None;
-
-        Self {
-            message_timestamp: Utc::now(),
-            date_created: existing.descriptor.date_created,
-            data: None,
-            data_format: existing.descriptor.data_format.clone(),
-            existing: Some(existing),
-            origin: Existing,
-            signer: Unsigned,
-            attesters: Unattested,
-            recipient: None,
-            protocol: None,
-            schema: None,
-            tags: None,
-            record_id: None,
-            published: None,
-            date_published: None,
-            protocol_role: None,
-            permission_grant_id: None,
-            delegated_grant: None,
-            encryption: None,
+        if let Some(protocol_path) = &self.protocol_path {
+            return Some(("protocolPath".to_string(), protocol_path.clone()));
         }
+        if let Some(context_id) = &self.context_id {
+            return Some(("contextId".to_string(), context_id.clone()));
+        }
+        if let Some(parent_id) = &self.parent_id {
+            return Some(("parentId".to_string(), parent_id.clone()));
+        }
+        if let Some(schema) = &self.schema {
+            return Some(("schema".to_string(), schema.clone()));
+        }
+        None
     }
 }
 
-/// State: New, Unattested, Unencrypted, and Unsigned.
-///
-/// Immutable properties are able be set.
-impl<'a> WriteBuilder<'a, New, Unattested, Unsigned> {
-    /// Set a protocol for the record.
+/// Implement  builder-like behaviour.
+impl RecordsFilter {
+    /// Add one or more authors to the filter.
     #[must_use]
-    pub fn protocol(mut self, protocol: ProtocolBuilder<'a>) -> Self {
-        self.protocol = Some(protocol);
+    pub fn add_author(mut self, author: impl Into<String>) -> Self {
+        match &mut self.author {
+            Some(OneOrMany::Many(existing)) => {
+                existing.push(author.into());
+            }
+            Some(OneOrMany::One(existing)) => {
+                self.author = Some(OneOrMany::Many(vec![existing.clone(), author.into()]));
+            }
+            None => {
+                self.author = Some(OneOrMany::One(author.into()));
+            }
+        }
         self
     }
 
-    /// Specify a schema to use with the record.
+    /// Add an attester to the filter.
+    #[must_use]
+    pub fn attester(mut self, attester: impl Into<String>) -> Self {
+        self.attester = Some(attester.into());
+        self
+    }
+
+    /// Add one or more recipients to the filter.
+    #[must_use]
+    pub fn add_recipient(mut self, recipient: impl Into<String>) -> Self {
+        match &mut self.recipient {
+            Some(OneOrMany::Many(existing)) => {
+                existing.push(recipient.into());
+            }
+            Some(OneOrMany::One(existing)) => {
+                self.recipient = Some(OneOrMany::Many(vec![existing.clone(), recipient.into()]));
+            }
+            None => {
+                self.recipient = Some(OneOrMany::One(recipient.into()));
+            }
+        }
+        self
+    }
+
+    /// Add a protocol to the filter.
+    #[must_use]
+    pub fn protocol(mut self, protocol: impl Into<String>) -> Self {
+        self.protocol = Some(protocol.into());
+        self
+    }
+
+    /// Add a protocol path to the filter.
+    #[must_use]
+    pub fn protocol_path(mut self, protocol_path: impl Into<String>) -> Self {
+        self.protocol_path = Some(protocol_path.into());
+        self
+    }
+
+    /// Specify a protocol schema on the filter.
     #[must_use]
     pub fn schema(mut self, schema: impl Into<String>) -> Self {
         self.schema = Some(schema.into());
         self
     }
 
-    /// Specify the write record's recipient .
-    #[must_use]
-    pub fn recipient(mut self, recipient: impl Into<String>) -> Self {
-        self.recipient = Some(recipient.into());
-        self
-    }
-}
-
-/// State: Unattested, and Unsigned.
-///
-///  Mutable properties properties are able to be set for both new and existing
-/// `Write` records.
-impl<O> WriteBuilder<'_, O, Unattested, Unsigned> {
-    /// Entry data as a CID or raw bytes.
-    #[must_use]
-    pub fn data(mut self, data: Data) -> Self {
-        self.data = Some(data);
-        self
-    }
-
-    /// The record's MIME type. Defaults to `application/json`.
-    #[must_use]
-    pub fn data_format(mut self, data_format: impl Into<String>) -> Self {
-        self.data_format = data_format.into();
-        self
-    }
-
-    /// Specify an ID to use for the permission request.
-    #[must_use]
-    pub fn record_id(mut self, record_id: impl Into<String>) -> Self {
-        self.record_id = Some(record_id.into());
-        self
-    }
-
-    /// Add a tag to the record.
-    #[must_use]
-    pub fn add_tag(mut self, name: impl Into<String>, tag: Tag) -> Self {
-        self.tags.get_or_insert_with(HashMap::new).insert(name.into(), tag);
-        self
-    }
-
-    /// Whether the record is published.
+    /// Add a published flag to the filter.
     #[must_use]
     pub const fn published(mut self, published: bool) -> Self {
         self.published = Some(published);
         self
     }
 
-    /// Specify a protocol role for the record.
+    /// Add a context ID to the filter.
     #[must_use]
-    pub fn protocol_role(mut self, protocol_role: impl Into<String>) -> Self {
-        self.protocol_role = Some(protocol_role.into());
+    pub fn context_id(mut self, context_id: impl Into<String>) -> Self {
+        self.context_id = Some(context_id.into());
         self
     }
 
-    /// Specifies the permission grant ID.
+    /// Add a record ID to the filter.
     #[must_use]
-    pub fn permission_grant_id(mut self, permission_grant_id: impl Into<String>) -> Self {
-        self.permission_grant_id = Some(permission_grant_id.into());
+    pub fn record_id(mut self, record_id: impl Into<String>) -> Self {
+        self.record_id = Some(record_id.into());
         self
     }
 
-    /// The delegated grant used with this record.
+    /// Add a parent ID to the filter.
     #[must_use]
-    pub fn delegated_grant(mut self, delegated_grant: DelegatedGrant) -> Self {
-        self.delegated_grant = Some(delegated_grant);
+    pub fn parent_id(mut self, parent_id: impl Into<String>) -> Self {
+        self.parent_id = Some(parent_id.into());
         self
     }
 
-    /// The encryption properties for the record.
+    /// Add a tag to the filter.
     #[must_use]
-    pub fn encryption(mut self, encryption: EncryptionProperty) -> Self {
-        self.encryption = Some(encryption);
+    pub fn add_tag(mut self, key: impl Into<String>, value: TagFilter) -> Self {
+        if let Some(existing) = &mut self.tags {
+            existing.insert(key.into(), value);
+        } else {
+            let mut tags = BTreeMap::new();
+            tags.insert(key.into(), value);
+            self.tags = Some(tags);
+        }
         self
     }
 
-    // ----------------------------------------------------------------
-    // Methods enabled soley for testing
-    // ----------------------------------------------------------------
-    /// Override message timestamp.
-    #[cfg(debug_assertions)]
+    /// Add a data format to the filter.
     #[must_use]
-    pub const fn message_timestamp(mut self, message_timestamp: DateTime<Utc>) -> Self {
-        self.message_timestamp = message_timestamp;
+    pub fn data_format(mut self, data_format: impl Into<String>) -> Self {
+        self.data_format = Some(data_format.into());
         self
     }
 
-    /// Override date created.
-    #[cfg(debug_assertions)]
+    /// Add a data size to the filter.
     #[must_use]
-    pub const fn date_created(mut self, date_created: DateTime<Utc>) -> Self {
-        self.date_created = date_created;
+    pub const fn data_size(mut self, data_size: Range<usize>) -> Self {
+        self.data_size = Some(data_size);
         self
     }
 
-    /// Override date published.
-    #[cfg(debug_assertions)]
+    /// Add a data CID to the filter.
     #[must_use]
-    pub const fn date_published(mut self, date_published: DateTime<Utc>) -> Self {
+    pub fn data_cid(mut self, data_cid: impl Into<String>) -> Self {
+        self.data_cid = Some(data_cid.into());
+        self
+    }
+
+    /// Add a date created to the filter.
+    #[must_use]
+    pub const fn date_created(mut self, date_created: DateRange) -> Self {
+        self.date_created = Some(date_created);
+        self
+    }
+
+    /// Add a date published to the filter.
+    #[must_use]
+    pub const fn date_published(mut self, date_published: DateRange) -> Self {
         self.date_published = Some(date_published);
         self
     }
+
+    /// Add a date updated to the filter.
+    #[must_use]
+    pub const fn date_updated(mut self, date_updated: DateRange) -> Self {
+        self.date_updated = Some(date_updated);
+        self
+    }
 }
 
-/// State: Unencrypted and Unsigned.
-impl<'a, O, A> WriteBuilder<'a, O, A, Unsigned> {
-    /// Logically (from user POV), have an attester sign the record.
-    ///
-    /// At this point, the builder simply captures the attester for use in the
-    /// final build step. Can only be done if the content hasn't been signed
-    /// or encrypted.
+/// Specifies the way that `RecordsQuery`results should be sorted.
+#[derive(Clone, Debug, Default, Deserialize, Serialize, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub enum Sort {
+    /// Sort `date_created` from oldest to newest.
+    #[serde(rename = "createdAscending")]
+    CreatedAsc,
+
+    /// Sort `date_created` newest to oldest.
+    #[serde(rename = "createdDescending")]
+    CreatedDesc,
+
+    /// Sort `date_published` from oldest to newest.
+    #[serde(rename = "publishedAscending")]
+    PublishedAsc,
+
+    /// Sort `date_published` from newest to oldest.
+    #[serde(rename = "publishedDescending")]
+    PublishedDesc,
+
+    /// Sort `message_timestamp` from oldest to newest.
+    #[serde(rename = "timestampAscending")]
+    #[default]
+    TimestampAsc,
+
+    /// Sort `message_timestamp` from newest to oldest.
+    #[serde(rename = "timestampDescending")]
+    TimestampDesc,
+}
+
+impl Sort {
+    /// Short-circuit testing for ascending/descending sort.
     #[must_use]
-    pub fn attest<S: Signer>(
-        self, attesters: &'a [&'a S],
-    ) -> WriteBuilder<'a, O, Attested<'a, S>, Unsigned> {
-        WriteBuilder {
-            attesters: Attested(attesters),
-            message_timestamp: self.message_timestamp,
-            recipient: self.recipient,
-            protocol: self.protocol,
-            schema: self.schema,
-            tags: self.tags,
-            record_id: self.record_id,
-            data: self.data,
-            data_format: self.data_format,
-            date_created: self.date_created,
-            published: self.published,
-            date_published: self.date_published,
-            protocol_role: self.protocol_role,
-            permission_grant_id: self.permission_grant_id,
-            delegated_grant: self.delegated_grant,
-            encryption: self.encryption,
-            existing: self.existing,
-            origin: self.origin,
-            signer: self.signer,
+    pub const fn is_ascending(&self) -> bool {
+        matches!(self, Self::CreatedAsc | Self::PublishedAsc | Self::TimestampAsc)
+    }
+}
+
+impl Display for Sort {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Self::CreatedAsc | Self::CreatedDesc => write!(f, "dateCreated"),
+            Self::PublishedAsc | Self::PublishedDesc => write!(f, "datePublished"),
+            Self::TimestampAsc | Self::TimestampDesc => write!(f, "messageTimestamp"),
         }
     }
 }
 
-// State: Unsigned
-impl<'a, O, A> WriteBuilder<'a, O, A, Unsigned> {
-    /// Logically (from user POV), sign the record.
-    ///
-    /// At this point, the builder simply captures the signer for use in the final
-    /// build step. Can only be done if the content hasn't been signed yet.
-    #[must_use]
-    pub fn sign(self, signer: &'a impl Signer) -> WriteBuilder<'a, O, A, Signed<'a, impl Signer>> {
-        WriteBuilder {
-            signer: Signed(signer),
+/// A tag filter is used when filter records by tag.
+#[derive(Clone, Debug, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub enum TagFilter {
+    /// Match tags starting with a string value.
+    StartsWith(String),
 
-            message_timestamp: self.message_timestamp,
-            recipient: self.recipient,
-            protocol: self.protocol,
-            schema: self.schema,
-            tags: self.tags,
-            record_id: self.record_id,
-            data: self.data,
-            data_format: self.data_format,
-            date_created: self.date_created,
-            published: self.published,
-            date_published: self.date_published,
-            protocol_role: self.protocol_role,
-            permission_grant_id: self.permission_grant_id,
-            delegated_grant: self.delegated_grant,
-            encryption: self.encryption,
-            existing: self.existing,
-            origin: self.origin,
-            attesters: self.attesters,
+    /// Filter tags by range.
+    Range(Range<usize>),
+
+    /// Filter by a specific value.
+    Equal(Tag),
+}
+
+impl Default for TagFilter {
+    fn default() -> Self {
+        Self::Equal(Tag::Empty)
+    }
+}
+
+/// The [`Read`] message expected by the handler.
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct Read {
+    /// Read descriptor.
+    pub descriptor: ReadDescriptor,
+
+    /// Message authorization.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub authorization: Option<Authorization>,
+}
+
+/// The [`Read`]  message descriptor.
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct ReadDescriptor {
+    /// The base descriptor
+    #[serde(flatten)]
+    pub base: Descriptor,
+
+    /// Defines the filter for the read.
+    pub filter: RecordsFilter,
+}
+
+/// [`ReadReply`] is returned by the handler in the [`Reply`] `body` field.
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct ReadReply {
+    /// The read reply entry.
+    pub entry: ReadReplyEntry,
+}
+
+/// [`ReadReplyEntry`] represents the [`Write`] entry returned for a successful
+/// 'read'.
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct ReadReplyEntry {
+    /// The latest `RecordsWrite` message of the record if record exists
+    /// (not deleted).
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub records_write: Option<Write>,
+
+    /// The `RecordsDelete` if the record is deleted.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub records_delete: Option<Delete>,
+
+    /// The initial write of the record if the returned `RecordsWrite` message
+    /// itself is not the initial write or if a `RecordsDelete` is returned.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub initial_write: Option<Write>,
+
+    /// The data for the record.
+    #[serde(skip)]
+    pub data: Option<io::Cursor<Vec<u8>>>,
+}
+
+/// The [`Subscribe`] message expected by the handler.
+#[derive(Clone, Debug, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct Subscribe {
+    /// The Subscribe descriptor.
+    pub descriptor: SubscribeDescriptor,
+
+    /// The message authorization.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub authorization: Option<Authorization>,
+}
+
+/// The [`Subscribe`]  message descriptor.
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct SubscribeDescriptor {
+    /// The base descriptor
+    #[serde(flatten)]
+    pub base: Descriptor,
+
+    /// Filter Records for subscribe.
+    pub filter: RecordsFilter,
+}
+
+/// [`SubscribeReply`] is returned by the handler in the [`Reply`] `body` field.
+#[derive(Debug, Default, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct SubscribeReply {
+    /// The subscription to the requested events.
+    /// N.B. serialization/deserialization is skipped because the subscriber
+    /// `Stream` is not serializable.
+    #[serde(skip)]
+    pub subscription: Subscriber,
+}
+
+/// The [`Write`] message expected by the handler.
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct Write {
+    /// Write descriptor.
+    pub descriptor: WriteDescriptor,
+
+    /// The message authorization.
+    pub authorization: Authorization,
+
+    /// The Entry CID for the record.
+    pub record_id: String,
+
+    /// Entry context.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub context_id: Option<String>,
+
+    /// Entry attestation.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub attestation: Option<Jws>,
+
+    /// Entry encryption.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub encryption: Option<EncryptionProperty>,
+
+    /// The base64url encoded data of the record if the data associated with
+    /// the record is equal or smaller than `MAX_ENCODING_SIZE`.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub encoded_data: Option<String>,
+
+    /// The data stream of the record if the data associated with the record
+    #[serde(skip)]
+    pub data_stream: Option<io::Cursor<Vec<u8>>>,
+}
+
+/// Signature payload.
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct SignaturePayload {
+    /// The standard signature payload.
+    #[serde(flatten)]
+    pub base: JwsPayload,
+
+    /// The ID of the record being signed.
+    pub record_id: String,
+
+    /// The context ID of the record being signed.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub context_id: Option<String>,
+
+    /// Attestation CID.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub attestation_cid: Option<String>,
+
+    /// Encryption CID .
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub encryption_cid: Option<String>,
+}
+
+/// Attestation payload.
+#[derive(Default, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct Attestation {
+    /// The attestation's descriptor CID.
+    pub descriptor_cid: String,
+}
+
+/// Delegated Grant is a special case of [`Write`] used in [`Authorization`]
+/// and [`Attestation`] grant references.
+///
+/// It is structured to cope with recursive references to [`Authorization`].
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct DelegatedGrant {
+    /// The grant's descriptor.
+    pub descriptor: WriteDescriptor,
+
+    ///The grant's authorization.
+    pub authorization: Box<Authorization>,
+
+    /// CID referencing the record associated with the message.
+    pub record_id: String,
+
+    /// Context id.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub context_id: Option<String>,
+
+    /// Encoded grant data.
+    pub encoded_data: String,
+}
+
+/// The [`Write`]  message descriptor.
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct WriteDescriptor {
+    /// The base descriptor
+    #[serde(flatten)]
+    pub base: Descriptor,
+
+    /// Entry's protocol.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub protocol: Option<String>,
+
+    /// The protocol path.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub protocol_path: Option<String>,
+
+    /// The record's recipient.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub recipient: Option<String>,
+
+    /// The record's schema.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub schema: Option<String>,
+
+    /// Tags associated with the record
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub tags: Option<HashMap<String, Tag>>,
+
+    /// The CID of the record's parent (if exists).
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub parent_id: Option<String>,
+
+    /// CID of the record's data.
+    pub data_cid: String,
+
+    /// The record's size in bytes.
+    pub data_size: usize,
+
+    /// The record's MIME type. For example, `application/json`.
+    pub data_format: String,
+
+    /// The datatime the record was created.
+    #[serde(serialize_with = "rfc3339_micros")]
+    pub date_created: DateTime<Utc>,
+
+    /// Indicates whether the record is published.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub published: Option<bool>,
+
+    /// The datetime of publishing, if published.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[serde(serialize_with = "rfc3339_micros_opt")]
+    pub date_published: Option<DateTime<Utc>>,
+}
+
+/// Tag value types.
+#[derive(Clone, Debug, Default, Deserialize, Serialize, PartialEq, Eq)]
+// #[serde(rename_all = "camelCase")]
+#[serde(untagged)]
+pub enum Tag {
+    /// Empty tag value.
+    #[default]
+    Empty,
+
+    /// String tag value.
+    String(String),
+
+    /// Number tag value.
+    Number(u64),
+
+    /// Boolean tag value.
+    Boolean(bool),
+}
+
+impl Tag {
+    /// Attempt to convert the tag value to a string.
+    #[must_use]
+    pub fn as_str(&self) -> Option<&str> {
+        match self {
+            Self::String(s) => Some(s.as_str()),
+            _ => None,
+        }
+    }
+
+    /// Attempt to convert the tag value to a u64.
+    #[must_use]
+    pub const fn as_u64(&self) -> Option<u64> {
+        match self {
+            Self::Number(n) => Some(*n),
+            _ => None,
+        }
+    }
+
+    /// Attempt to convert the tag value to a bool.
+    #[must_use]
+    pub const fn as_bool(&self) -> Option<bool> {
+        match self {
+            Self::Boolean(b) => Some(*b),
+            _ => None,
         }
     }
 }
 
-// State: Signed.
+/// For consistency, [`WriteReply`] is returned by the handler in the [`Reply`]
+/// `body` field, but contains no data.
+#[derive(Clone, Debug)]
+pub struct WriteReply;
 
-/// Builder is ready to build once the `sign` step is complete (i.e. the Signer
-/// is set).
-impl<O, A, S: Signer> WriteBuilder<'_, O, A, Signed<'_, S>> {
-    // TODO: break into separate functions
-    fn to_write(&self, author_did: &str) -> Result<Write> {
-        let mut write = if let Some(write) = &self.existing {
-            write.clone()
-        } else {
-            // set immutable properties
-            let mut write = Write {
-                descriptor: WriteDescriptor {
-                    base: Descriptor {
-                        interface: Interface::Records,
-                        method: Method::Write,
-                        ..Descriptor::default()
-                    },
-                    date_created: self.date_created,
-                    recipient: self.recipient.clone(),
-                    ..WriteDescriptor::default()
-                },
-                ..Write::default()
+/// Encryption settings.
+#[derive(Clone, Debug, Default)]
+pub struct EncryptOptions<'a> {
+    /// The algorithm to use to encrypt the message data.
+    content_algorithm: ContentAlgorithm,
+
+    /// The algorithm to use to encrypt (or derive) the content encryption key
+    /// (CEK).
+    key_algorithm: KeyAlgorithm,
+
+    /// The data to encrypt.
+    data: &'a [u8],
+
+    /// An array of inputs specifying how the CEK key is to be encrypted. Each
+    /// entry in the array will result in a unique ciphertext for the CEK.
+    recipients: Vec<Recipient>,
+}
+
+/// Encrypted data. Intermediate work product.
+#[derive(Zeroize, ZeroizeOnDrop)]
+pub struct Encrypted {
+    /// The algorithm to use to encrypt the message data.
+    #[zeroize(skip)]
+    content_algorithm: ContentAlgorithm,
+
+    /// The algorithm to use to encrypt (or derive) the content encryption key
+    /// (CEK).
+    #[zeroize(skip)]
+    key_algorithm: KeyAlgorithm,
+
+    /// An array of inputs specifying how the CEK key is to be encrypted. Each
+    /// entry in the array will result in a unique ciphertext for the CEK.
+    #[zeroize(skip)]
+    pub recipients: Vec<Recipient>,
+
+    /// The content encryption key (CEK) used to encrypt the data.
+    pub cek: Vec<u8>,
+
+    /// The initialization vector (IV) used to encrypt the data.
+    pub iv: String,
+
+    /// The additional authenticated data (AAD) used to encrypt the data.
+    pub tag: String,
+
+    /// The ciphertext.
+    pub ciphertext: Vec<u8>,
+}
+
+/// Encryption key settings.
+#[derive(Clone, Debug, Default)]
+pub struct Recipient {
+    /// The identifier of the recipient's public key used to encrypt the
+    /// content encryption key (CEK).
+    pub key_id: String,
+
+    /// The recipient's public key used to encrypt the CEK.
+    pub public_key: PublicKeyJwk,
+
+    /// The content encryption key (CEK) derivation scheme.
+    pub derivation_scheme: DerivationScheme,
+}
+
+impl<'a> EncryptOptions<'a> {
+    /// Create a new `EncryptionOptions`.
+    #[must_use]
+    pub const fn new() -> Self {
+        Self {
+            content_algorithm: ContentAlgorithm::A256Gcm,
+            key_algorithm: KeyAlgorithm::EcdhEsA256Kw,
+            data: &[],
+            recipients: vec![],
+        }
+    }
+
+    /// Set the content encryption algorithm.
+    #[must_use]
+    pub const fn content_algorithm(mut self, algorithm: ContentAlgorithm) -> Self {
+        self.content_algorithm = algorithm;
+        self
+    }
+
+    /// Set the key encryption algorithm.
+    #[must_use]
+    pub const fn key_algorithm(mut self, algorithm: KeyAlgorithm) -> Self {
+        self.key_algorithm = algorithm;
+        self
+    }
+
+    /// Set the data to encrypt.
+    #[must_use]
+    pub const fn data(mut self, data: &'a [u8]) -> Self {
+        self.data = data;
+        self
+    }
+
+    /// Add a recipient to the encryption options.
+    #[must_use]
+    pub fn with_recipient(mut self, recipient: Recipient) -> Self {
+        self.recipients.push(recipient);
+        self
+    }
+
+    /// Encrypt the data using the specified encryption options, retaining the
+    /// CEK, IV, and AAD tag for later use.
+    ///
+    /// # Errors
+    ///
+    /// Will fail if the [`Protected`] struct cannot be serialized to JSON or
+    /// if the provided data cannot be encrypted using the specified content
+    /// encryption algorithm.
+    pub fn encrypt(&mut self) -> Result<Encrypted> {
+        use aes_gcm::Aes256Gcm;
+        use aes_gcm::aead::KeyInit;
+
+        let cek = Aes256Gcm::generate_key(&mut rand::thread_rng());
+        let protected = Protected {
+            enc: self.content_algorithm.clone(),
+            alg: None,
+        };
+        let aad = serde_json::to_vec(&protected)?;
+
+        let encrypted = match self.content_algorithm {
+            ContentAlgorithm::A256Gcm => jwe::a256gcm(self.data, &cek.into(), &aad)?,
+            ContentAlgorithm::XChaCha20Poly1305 => {
+                jwe::xchacha20_poly1305(self.data, &cek.into(), &aad)?
+            }
+        };
+
+        Ok(Encrypted {
+            content_algorithm: self.content_algorithm.clone(),
+            key_algorithm: self.key_algorithm.clone(),
+            recipients: self.recipients.clone(),
+            cek: cek.to_vec(),
+            iv: encrypted.iv,
+            tag: encrypted.tag,
+            ciphertext: encrypted.ciphertext,
+        })
+    }
+}
+
+impl Encrypted {
+    /// Add a recipient to the encryption options.
+    #[must_use]
+    pub fn add_recipient(mut self, recipient: Recipient) -> Self {
+        self.recipients.push(recipient);
+        self
+    }
+
+    /// Finalize the encryption process, returning the encryption properties
+    ///
+    /// # Errors
+    ///
+    /// Will fail if the CEK cannot be encrypted. 
+    pub fn finalize(self) -> Result<EncryptionProperty> {
+        // encryption property
+        let mut encryption = EncryptionProperty {
+            algorithm: self.content_algorithm.clone(),
+            initialization_vector: self.iv.clone(),
+            message_authentication_code: Some(self.tag.clone()),
+            key_encryption: vec![],
+        };
+
+        // add `EncryptedKey` for each recipient
+        for recipient in &self.recipients {
+            // recipient's public key
+            let jwk = &recipient.public_key;
+            let decoded = if jwk.crv == Curve::Ed25519 {
+                Base64UrlUnpadded::decode_vec(&jwk.x)?
+            } else {
+                let mut decoded = Base64UrlUnpadded::decode_vec(&jwk.x)?;
+                let Some(y) = &jwk.y else {
+                    return Err(unexpected!("missing y"));
+                };
+                decoded.extend(&Base64UrlUnpadded::decode_vec(y)?);
+                decoded
             };
 
-            if let Some(record_id) = self.record_id.clone() {
-                write.record_id = record_id;
-            }
-            if let Some(settings) = self.protocol.clone() {
-                let normalized = utils::uri::clean(settings.protocol)?;
-                write.descriptor.protocol = Some(normalized);
-                write.descriptor.protocol_path = Some(settings.protocol_path.to_string());
+            // create `jwe::Recipient` for call to jwe key wrapping function
+            let recip = jwe::Recipient {
+                key_id: recipient.key_id.clone(),
+                public_key: jwe::PublicKey::try_from(decoded)?,
+            };
+            let cek: [u8; 32] =
+                self.cek.clone().try_into().map_err(|_| unexpected!("invalid CEK key"))?;
 
-                // parent_id == last segment of  `parent_context_id`
-                if let Some(parent_context_id) = &settings.parent_context_id {
-                    write.descriptor.parent_id =
-                        parent_context_id.split('/').next_back().map(ToString::to_string);
+            // encrypt cek
+            let ke = match self.key_algorithm {
+                KeyAlgorithm::EcdhEsA256Kw => jwe::ecdh_a256kw(&cek, &recip)?,
+                KeyAlgorithm::EciesEs256K => jwe::ecies_es256k(&cek, &recip)?,
+                KeyAlgorithm::EcdhEs => {
+                    return Err(unexpected!("ECDH-ES requires a single recipient"));
                 }
+            };
+
+            // unpack `jwe::KeyEncryption` into `EncryptedKey`
+            let mut encrypted = EncryptedKey {
+                root_key_id: recipient.key_id.clone(),
+                algorithm: ke.header.alg.clone(),
+                ephemeral_public_key: ke.header.epk.clone(),
+                initialization_vector: ke.header.iv.clone(),
+                message_authentication_code: ke.header.tag.clone(),
+                cek: ke.encrypted_key.clone(),
+                derivation_scheme: recipient.derivation_scheme.clone(),
+                derived_public_key: None,
+            };
+
+            // attach the public key when derivation scheme is protocol-context,
+            // so that the responder to this message is able to encrypt the
+            // content encryption key using the same protocol-context derived
+            // public key, without needing the knowledge of the corresponding
+            // private key
+            if recipient.derivation_scheme == DerivationScheme::ProtocolContext {
+                encrypted.derived_public_key = Some(recipient.public_key.clone());
             }
-            if let Some(s) = &self.schema {
-                write.descriptor.schema = Some(utils::uri::clean(s)?);
-            }
 
-            write
-        };
-
-        // mutable properties
-        write.descriptor.base.message_timestamp = self.message_timestamp;
-        write.descriptor.data_format.clone_from(&self.data_format);
-
-        // tags
-        if let Some(tags) = self.tags.clone() {
-            write.descriptor.tags = Some(tags);
+            encryption.key_encryption.push(encrypted);
         }
 
-        // published state
-        if let Some(published) = self.published {
-            write.descriptor.published = Some(published);
-
-            // set/unset `date_published`
-            if published {
-                write.descriptor.date_published =
-                    Some(self.date_published.unwrap_or(self.message_timestamp));
-            } else {
-                write.descriptor.date_published = None;
-            }
-        }
-
-        match &self.data {
-            Some(Data::Stream(stream)) => {
-                let (data_cid, data_size) = cid::from_reader(stream.clone())?;
-                write.descriptor.data_cid = data_cid;
-                write.descriptor.data_size = data_size;
-                write.data_stream = Some(stream.clone());
-            }
-            Some(Data::Bytes(data)) => {
-                // calculate CID and size only — don't add to `data_stream`
-                let data_cid = cid::from_value(data)?;
-                write.descriptor.data_cid = data_cid;
-                write.descriptor.data_size = data.len();
-            }
-            Some(Data::Cid { data_cid, data_size }) => {
-                write.descriptor.data_cid.clone_from(data_cid);
-                write.descriptor.data_size = *data_size;
-            }
-            None => {}
-        }
-
-        if let Some(encryption) = &self.encryption {
-            for key in &encryption.key_encryption {
-                if key.derivation_scheme == DerivationScheme::ProtocolPath
-                    && self.protocol.is_none()
-                {
-                    return Err(anyhow!(
-                        "`protocol` must be specified when using `protocols` encryption scheme"
-                    ));
-                }
-                if key.derivation_scheme == DerivationScheme::Schemas && self.schema.is_none() {
-                    return Err(anyhow!(
-                        "`schema` must be specified when using `schema` encryption scheme"
-                    ));
-                }
-            }
-
-            write.encryption = Some(encryption.clone());
-        }
-
-        write.authorization = Authorization {
-            author_delegated_grant: self.delegated_grant.clone(),
-            ..Authorization::default()
-        };
-
-        // compute `record_id` when not provided
-        if write.record_id.is_empty() {
-            write.record_id = write.entry_id(author_did)?;
-        }
-
-        // compute `context_id` if this is a protocol-space record
-        if let Some(settings) = &self.protocol {
-            if let Some(parent_context_id) = &write.context_id {
-                write.context_id = Some(format!("{parent_context_id}/{}", write.record_id));
-            } else if let Some(parent_context_id) = &settings.parent_context_id {
-                write.context_id = Some(format!("{parent_context_id}/{}", write.record_id));
-            } else {
-                write.context_id = Some(write.record_id.clone());
-            }
-        }
-
-        Ok(write)
+        Ok(encryption)
     }
 }
 
-impl<O, A: Signer, S: Signer> WriteBuilder<'_, O, Attested<'_, A>, Signed<'_, S>> {
-    async fn attestation(self, descriptor: &WriteDescriptor) -> Result<Jws> {
-        let payload = Attestation {
-            descriptor_cid: cid::from_value(descriptor)?,
-        };
-        let Some(attester) = self.attesters.0.first() else {
-            return Err(anyhow!("attesters is empty"));
-        };
-        JwsBuilder::new().payload(payload).add_signer(*attester).build().await
-    }
+/// `EncryptionProperty` contains information about the encryption used when
+/// encrypting a `Write` message. The information is used by the recipient
+/// to decrypt the message.
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct EncryptionProperty {
+    /// The algorithm used to encrypt the data. Equivalent to the JWE Encryption
+    /// Algorithm (JWE header `enc` property).
+    pub algorithm: ContentAlgorithm,
+
+    /// The initialization vector used to encrypt the data.
+    pub initialization_vector: String,
+
+    /// One or more objects, each containing information about the
+    /// Content Encryption Key (CEK) used to encrypt the data.
+    pub key_encryption: Vec<EncryptedKey>,
+
+    /// The message authentication code.
+    /// Equivalent to the JWE Authentication Tag (JWE `tag` property).
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub message_authentication_code: Option<String>,
 }
 
-/// State: Unattested, Unencrypted, and Signed.
-impl<O, S: Signer> WriteBuilder<'_, O, Unattested, Signed<'_, S>> {
-    /// Build the `Write` message.
-    ///
-    /// # Errors
-    ///
-    /// This method will fail when there is an issue authorizing the message.
-    pub async fn build(self) -> Result<Write> {
-        let author_did = if let Some(grant) = &self.delegated_grant {
-            grant.authorization.signature.did()?
-        } else {
-            did_from_kid(&self.signer.0.verification_method().await?)?
-        };
+/// The encrypted Content Encryption Key (CEK). Equivalent to the JWE
+/// Encrypted Key (JWE `encrypted_key` property), this is the key used to
+/// encrypt the data.
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct EncryptedKey {
+    /// The fully qualified key ID (e.g. did:example:abc#encryption-key-id)
+    /// of the root public key used to encrypt the symmetric encryption key.
+    pub root_key_id: String,
 
-        let mut write = self.to_write(&author_did)?;
-        write.sign_as_author(self.permission_grant_id, self.protocol_role, self.signer.0).await?;
-        Ok(write)
-    }
-}
+    /// The derived public key.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub derived_public_key: Option<PublicKeyJwk>,
 
-/// State: Attested, and Signed.
-impl<'a, O, A: Signer, S: Signer> WriteBuilder<'a, O, Attested<'a, A>, Signed<'a, S>> {
-    /// Build the `Write` message.
-    ///
-    /// # Errors
-    ///
-    /// This method will fail when there is an issue attesting to or
-    /// authorizing the message.
-    pub async fn build(self) -> Result<Write> {
-        let author_did = if let Some(grant) = &self.delegated_grant {
-            grant.authorization.signature.did()?
-        } else {
-            did_from_kid(&self.signer.0.verification_method().await?)?
-        };
+    /// The content encryption key (CEK) derivation scheme.
+    pub derivation_scheme: DerivationScheme,
 
-        let signer = self.signer.0;
-        let protocol_role = self.protocol_role.clone();
-        let permission_grant_id = self.permission_grant_id.clone();
+    /// The algorithm used to encrypt the data. Equivalent to the JWE Encryption
+    /// Algorithm (JWE header `alg` property).
+    pub algorithm: KeyAlgorithm,
 
-        let mut write = self.to_write(&author_did)?;
-        write.attestation = Some(self.attestation(&write.descriptor).await?);
-        write.sign_as_author(permission_grant_id, protocol_role, signer).await?;
-        Ok(write)
-    }
-}
+    /// The ephemeral public key used to encrypt the data.
+    pub ephemeral_public_key: PublicKeyJwk,
 
-fn did_from_kid(kid: &str) -> Result<String> {
-    let parts: Vec<&str> = kid.split('#').collect();
-    if parts.len() != 2 {
-        return Err(anyhow!("Invalid key ID"));
-    }
-    Ok(parts[0].to_string())
+    /// The initialization vector used to encrypt the data.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub initialization_vector: Option<String>,
+
+    /// The message authentication code.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub message_authentication_code: Option<String>,
+
+    /// The encrypted Content Encryption Key (CEK). Equivalent to the JWE
+    /// Encrypted Key (JWE `encrypted_key` property), this is the key used to
+    /// encrypt the data.
+    #[serde(rename = "encryptedKey")]
+    pub cek: String,
 }
