+++ src/messages/subscribe.rs
@@ -5,15 +5,14 @@
 
 use futures::{StreamExt, future};
 use http::StatusCode;
-use serde::{Deserialize, Serialize};
 
 use crate::authorization::Authorization;
 use crate::endpoint::{Message, Reply, Status};
-use crate::event::{SubscribeFilter, Subscriber};
-use crate::messages::MessagesFilter;
+use crate::event::SubscribeFilter;
+use crate::interfaces::Descriptor;
+use crate::interfaces::messages::{Subscribe, SubscribeReply};
 use crate::provider::{EventStream, MessageStore, Provider};
-use crate::utils::cid;
-use crate::{Descriptor, Result, forbidden, grants};
+use crate::{Result, forbidden, grants};
 
 /// Handle — or process — a [`Subscribe`] message.
 ///
@@ -49,22 +48,8 @@ pub async fn handle(
     })
 }
 
-/// The [`Subscribe`] message expected by the handler.
-#[derive(Clone, Debug, Default, Deserialize, Serialize)]
-pub struct Subscribe {
-    /// The Subscribe descriptor.
-    pub descriptor: SubscribeDescriptor,
-
-    /// The message authorization.
-    pub authorization: Authorization,
-}
-
 impl Message for Subscribe {
     type Reply = SubscribeReply;
-
-    fn cid(&self) -> Result<String> {
-        cid::from_value(self)
-    }
 
     fn descriptor(&self) -> &Descriptor {
         &self.descriptor.base
@@ -109,24 +94,4 @@ impl Subscribe {
 
         Ok(())
     }
-}
-
-/// [`SubscribeReply`] is returned by the handler in the [`Reply`] `body` field.
-#[derive(Debug, Deserialize, Serialize)]
-pub struct SubscribeReply {
-    /// The subscription to the requested events.
-    #[serde(skip)]
-    pub subscription: Subscriber,
-}
-
-/// The [`Subscribe`]  message descriptor.
-#[derive(Clone, Debug, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SubscribeDescriptor {
-    /// The base descriptor
-    #[serde(flatten)]
-    pub base: Descriptor,
-
-    /// Filters to apply when subscribing to messages.
-    pub filters: Vec<MessagesFilter>,
 }
