+++ src/records/query.rs
@@ -5,16 +5,17 @@
 //! [`Delete`]) messages.
 
 use http::StatusCode;
-use serde::{Deserialize, Serialize};
 
 use crate::authorization::Authorization;
 use crate::endpoint::{Message, Reply, Status};
 use crate::grants::Grant;
+use crate::interfaces::Descriptor;
+use crate::interfaces::records::{Query, QueryReply, QueryReplyEntry, RecordsFilter, Sort, Write};
 use crate::provider::{MessageStore, Provider};
-use crate::records::{RecordsFilter, Write, protocol};
-use crate::store::{self, Cursor, Pagination, RecordsQueryBuilder, Sort};
+use crate::records::protocol;
+use crate::store::{self, RecordsQueryBuilder};
 use crate::utils::cid;
-use crate::{Descriptor, Result, forbidden, unexpected, utils};
+use crate::{Result, forbidden, unexpected, utils};
 
 /// Handle — or process — a [`Query`] message.
 ///
@@ -101,18 +102,6 @@ pub async fn handle(
     })
 }
 
-/// The [`Query`] message expected by the handler.
-#[derive(Clone, Debug, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct Query {
-    /// The Query descriptor.
-    pub descriptor: QueryDescriptor,
-
-    /// The message authorization.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub authorization: Option<Authorization>,
-}
-
 impl Message for Query {
     type Reply = QueryReply;
 
@@ -131,33 +120,6 @@ impl Message for Query {
     async fn handle(self, owner: &str, provider: &impl Provider) -> Result<Reply<Self::Reply>> {
         handle(owner, self, provider).await
     }
-}
-
-/// [`QueryReply`] is returned by the handler in the [`Reply`] `body` field.
-#[derive(Clone, Debug, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct QueryReply {
-    /// Query reply entries.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub entries: Option<Vec<QueryReplyEntry>>,
-
-    /// Pagination cursor.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub cursor: Option<Cursor>,
-}
-
-/// [`QueryReplyEntry`] represents a [`Write`] entry returned by the query.
-#[derive(Clone, Debug, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct QueryReplyEntry {
-    /// The `RecordsWrite` message of the record if record exists.
-    #[serde(flatten)]
-    pub write: Write,
-
-    /// The initial write of the record if the returned `RecordsWrite` message
-    /// itself is not the initial write.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub initial_write: Option<Write>,
 }
 
 impl Query {
@@ -282,24 +244,4 @@ impl Query {
 
         Ok(store_query.build())
     }
-}
-
-/// The [`Query`] message descriptor.
-#[derive(Clone, Debug, Default, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct QueryDescriptor {
-    /// The base descriptor
-    #[serde(flatten)]
-    pub base: Descriptor,
-
-    /// Filter Records for query.
-    pub filter: RecordsFilter,
-
-    /// Specifies how dates should be sorted.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub date_sort: Option<Sort>,
-
-    /// The pagination cursor.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub pagination: Option<Pagination>,
 }
