+++ tests/messages_subscribe.rs
@@ -6,10 +6,10 @@ use std::sync::LazyLock;
 use std::time::Duration;
 
 use dwn_node::authorization::Authorization;
-use dwn_node::interfaces::grants::{GrantBuilder, Scope};
-use dwn_node::interfaces::messages::{MessagesFilter, QueryBuilder, SubscribeBuilder};
-use dwn_node::interfaces::protocols::{ConfigureBuilder, Definition};
-use dwn_node::interfaces::records::{Data, ProtocolBuilder, WriteBuilder};
+use dwn_node::client::grants::{GrantBuilder, Scope};
+use dwn_node::client::messages::{MessagesFilter, QueryBuilder, SubscribeBuilder};
+use dwn_node::client::protocols::{ConfigureBuilder, Definition};
+use dwn_node::client::records::{Data, ProtocolBuilder, WriteBuilder};
 use dwn_node::{Error, Interface, Message, Method, StatusCode, endpoint};
 use futures::StreamExt;
 use test_node::keystore::{self, Keyring};
