+++ tests/protocols_configure.rs
@@ -6,8 +6,8 @@
 use std::collections::BTreeMap;
 use std::sync::LazyLock;
 
-use dwn_node::interfaces::grants::{GrantBuilder, RevocationBuilder, Scope};
-use dwn_node::interfaces::protocols::{
+use dwn_node::client::grants::{GrantBuilder, RevocationBuilder, Scope};
+use dwn_node::client::protocols::{
     Action, ActionRule, Actor, ConfigureBuilder, Definition, ProtocolType, QueryBuilder, RuleSet,
 };
 use dwn_node::provider::MessageStore;
