+++ tests/protocols_query.rs
@@ -9,8 +9,8 @@ use std::time::Duration;
 use base64ct::{Base64UrlUnpadded, Encoding};
 use chrono::Days;
 use credibil_infosec::jose::jws::{Jws, Protected, Signature};
-use dwn_node::interfaces::grants::{GrantBuilder, RevocationBuilder, Scope};
-use dwn_node::interfaces::protocols::{
+use dwn_node::client::grants::{GrantBuilder, RevocationBuilder, Scope};
+use dwn_node::client::protocols::{
     ConfigureBuilder, Definition, ProtocolType, ProtocolsFilter, QueryBuilder,
 };
 use dwn_node::{Error, Method, StatusCode, cid, endpoint};
