+++ tests/records_subscribe.rs
@@ -3,7 +3,7 @@
 use std::sync::LazyLock;
 use std::time::Duration;
 
-use dwn_node::interfaces::records::{
+use dwn_node::client::records::{
     Data, QueryBuilder, RecordsFilter, SubscribeBuilder, WriteBuilder,
 };
 use dwn_node::{Message, StatusCode, endpoint};
